+++ game/end_to_end_tests/test_play_through.py
@@ -143,7 +143,7 @@ class TestPlayThrough(BaseGameTest):
         self._complete_level(38)
 
     def test_level_039(self):
-        self._complete_level(39)
+        self._complete_level(39, check_route_score=False)
 
     def test_level_040(self):
         self._complete_level(40)